% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_convergence.R
\name{check_convergence}
\alias{check_convergence}
\title{Checking the algorithm converged}
\usage{
check_convergence(object)
}
\arguments{
\item{object}{Output object from the \code{\link{reslr_mcmc}}}
}
\value{
Statement of convergence
}
\description{
Checking the algorithm converged
}
\examples{
data <- NAACproxydata \%>\% dplyr::filter(Site == "Cedar Island")
input_data <- reslr_load(data = data)
jags_output <- reslr_mcmc(input_data = input_data, model_type = "eiv_slr_t")
check_convergence(object = jags_output)
}
