% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation_test.R
\name{cross_validation_test}
\alias{cross_validation_test}
\title{This function examines the number of knots used to create the spline model fit.}
\usage{
cross_validation_test(object, ...)
}
\arguments{
\item{object}{Output object from the \code{\link{reslr_mcmc}}}

\item{...}{Not in use}
}
\value{
A list containing convergence diagnostics and parameter estimates for the output.
}
\description{
This function examines the number of knots used to create the spline model fit.
}
\examples{
data <- NAACproxydata \%>\% dplyr::filter(Site == "Cedar Island")
input_data <- reslr_load(data = data)
jags_output <- reslr_mcmc(input_data = input_data, model_type = "ni_spline_t")
cross_validation_test(object = jags_output)
}
