% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_input_data.R
\name{load_input_data}
\alias{load_input_data}
\title{Loading input data}
\usage{
load_input_data(data, n_prediction = 100, tide_gauge_included = FALSE)
}
\arguments{
\item{data}{The data of interest}

\item{n_prediction}{Predictions over every 100 years(default) can vary based on user preference}

\item{tide_gauge_included}{Including decadaly average tide gauge data from PSMSL website}
}
\value{
A list containing data frame of data and prediction grid
}
\description{
Loading input data
}
\examples{
data <- NAACproxydata \%>\% dplyr::filter(Site == "Cedar Island")
load_input_data(data = data)
}
