% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_estimate.R
\name{parameter_estimate}
\alias{parameter_estimate}
\title{Examining the parameter estimates from JAGS}
\usage{
parameter_estimate(jags_output)
}
\arguments{
\item{jags_output}{The JAGS output}
}
\value{
Summary of the parameter estimates
}
\description{
Examining the parameter estimates from JAGS
}
\examples{
data <- NAACproxydata \%>\% dplyr::filter(Site == "Cedar Island")
input_data <- load_input_data(data = data)
parameter_estimate(jags_output = run_mcmc(input_data = input_data, model_type = "eiv_slr_t"))
}
