% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.reslr_output.R, R/plot_result.R
\name{plot.reslr_output}
\alias{plot.reslr_output}
\title{Plotting the results for Noisy Input GAM decomposition}
\usage{
\method{plot}{reslr_output}(jags_output, plot_tide_gauges = FALSE)

\method{plot}{reslr_output}(jags_output, plot_tide_gauges = FALSE)
}
\arguments{
\item{jags_output}{List of JAGS output}

\item{plot_tide_gauges}{Plotting the tide gauge data as well as proxy data}
}
\value{
Plot of model fit

Plot of model fit
}
\description{
Plotting the results for Noisy Input GAM decomposition

Plotting the results for Noisy Input GAM decomposition
}
\examples{
data <- NAACproxydata \%>\% dplyr::filter(Site == "Cedar Island")
input_data <- load_input_data(data = data)
jags_output <- run_mcmc(input_data, model_type = "eiv_slr_t")
plot(jags_output)
data <- NAACproxydata \%>\% dplyr::filter(Site == "Cedar Island")
input_data <- load_input_data(data = data)
jags_output <- run_mcmc(input_data, model_type = "eiv_slr_t")
plot_result(jags_output)
}
