% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.reslr_input.R
\name{print.reslr_input}
\alias{print.reslr_input}
\title{Print a reslr output object which is created from the \code{reslr_load} function.}
\usage{
\method{print}{reslr_input}(x, ...)
}
\arguments{
\item{x}{An object of class \code{reslr_input}}

\item{...}{Other arguments (not supported)}
}
\value{
A neat presentation of your input reslr object
}
\description{
In this function, the reslr input object is printed. This is a high-level summary which provides the number of observations and the number of sites utilised in the dataset.
}
\examples{
\dontrun{
data <- NAACproxydata \%>\% dplyr::filter(Site == "Cedar Island")
reslr_input <- reslr_load(data = data)
print(x = reslr_input)}
}
