% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.reslr_output.R
\name{print.reslr_output}
\alias{print.reslr_output}
\title{Print a reslr output object}
\usage{
\method{print}{reslr_output}(x, ...)
}
\arguments{
\item{x}{An object of class \code{reslr_output}}

\item{...}{Other arguments (not supported)}
}
\value{
Returns a neat summary of the object
}
\description{
Print a reslr output object
}
\examples{
data <- NAACproxydata \%>\% dplyr::filter(Site == "Cedar Island")
input_data <- reslr_load(data = data)
jags_output <- reslr_mcmc(input_data = input_data, model_type = "eiv_slr_t")
print(x = jags_output)
}
