% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reslr_load.R
\name{reslr_load}
\alias{reslr_load}
\title{Loading input data}
\usage{
reslr_load(
  data,
  n_prediction = 100,
  include_tide_gauge = FALSE,
  include_linear_rate = FALSE,
  input_Age_type = "CE"
)
}
\arguments{
\item{data}{The data of interest}

\item{n_prediction}{Predictions over every 100 years(default) can vary based on user preference}

\item{include_tide_gauge}{Including decadaly average tide gauge data from PSMSL website}

\item{include_linear_rate}{User decides to include linear_rate and linear_rate_err}

\item{input_Age_type}{The inputted age in years CE or year BCE}
}
\value{
A list containing data frame of data and prediction grid
}
\description{
Loading input data
}
\examples{
data <- NAACproxydata \%>\% dplyr::filter(Site == "Cedar Island")
reslr_load(data = data)
}
