% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mcmc.R
\name{run_mcmc}
\alias{run_mcmc}
\title{Noisy Input Generalised Additive Models}
\usage{
run_mcmc(
  input_data,
  model_type,
  n_cp = 1,
  igp_smooth = 0.2,
  n_iterations = 5000,
  n_burnin = 1000,
  n_thin = 4,
  n_chains = 3
)
}
\arguments{
\item{input_data}{Input data from the load_input_data function}

\item{model_type}{Statistical Model type}

\item{n_cp}{Number of change points 1,2 or 3}

\item{igp_smooth}{Informs prior for the smoothness (correlation) parameter if model = "igp" is chosen. Choose a value between 0 and 1. Closer to 1 will increase smoothness.}

\item{n_iterations}{Number of iterations}

\item{n_burnin}{Burnin value}

\item{n_thin}{Thinning value}

\item{n_chains}{Number of chains}
}
\value{
JAGS list output
}
\description{
Noisy Input Generalised Additive Models
}
\examples{
data <- NAACproxydata \%>\% dplyr::filter(Site == "Cedar Island")
input_data <- load_input_data(data = data)
run_mcmc(input_data = input_data, model_type = "eiv_slr_t")
}
