% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_convergence.R
\name{check_convergence}
\alias{check_convergence}
\title{Checking the algorithm converged}
\usage{
check_convergence(jags_output)
}
\arguments{
\item{jags_output}{Output from the JAGS}
}
\value{
Statement of convergence
}
\description{
Checking the algorithm converged
}
\examples{
data <- NAACproxydata \%>\% dplyr::filter(Site == "Cedar Island")
input_data <- reslr_load(data = data)
check_convergence(jags_output = reslr_mcmc(input_data = input_data, model_type = "eiv_slr_t"))
}
