% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.reslr_input.R
\name{plot.reslr_input}
\alias{plot.reslr_input}
\title{Plot data with measurement uncertainty}
\usage{
\method{plot}{reslr_input}(
  input_data,
  title = "Plot of the raw data",
  xlab = "Age (CE)",
  ylab = "Relative Sea Level (m)",
  plot_tide_gauges = FALSE
)
}
\arguments{
\item{input_data}{Input data from the load_input_data function}

\item{title}{Title of the Plot}

\item{xlab}{Labeling the x-axis}

\item{ylab}{Labeling the y-axis}

\item{plot_tide_gauges}{Plotting the tide gauge data with the proxy records}
}
\value{
Plot of the RSL data
}
\description{
Plot data with measurement uncertainty
}
\examples{
full_dataset <- load_input_data(NAACproxydata)
plot_data(input_data = full_dataset)
}
