% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.reslr_output.R
\name{plot.reslr_output}
\alias{plot.reslr_output}
\title{Plotting the results for Noisy Input GAM decomposition}
\usage{
\method{plot}{reslr_output}(x, plot_tide_gauges = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{reslr_output} and \code{model_type} created via \code{\link{run_mcmc}}}

\item{plot_tide_gauges}{Plotting the tide gauge data as well as proxy data}

\item{...}{Not used}
}
\value{
Plot of model fit
}
\description{
Plotting the results for Noisy Input GAM decomposition
}
\examples{
data <- NAACproxydata \%>\% dplyr::filter(Site == "Cedar Island")
x <- load_input_data(data = data)
jags_output <- run_mcmc(x, model_type = "eiv_slr_t")
plot(x =jags_output)
}
