% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.reslr_output.R
\name{plot.reslr_output}
\alias{plot.reslr_output}
\title{Plotting the results for each statistical model from the \code{reslr_mcmc} function.}
\usage{
\method{plot}{reslr_output}(x, plot_tide_gauges = FALSE, plot_type = c("model_fit_plot"), ...)
}
\arguments{
\item{x}{An object of class \code{reslr_output} and \code{model_type} created via \code{\link{reslr_mcmc}}}

\item{plot_tide_gauges}{Plotting the tide gauge data as well as proxy data}

\item{plot_type}{The user can select the type of output plot they require from the following: "rate_plot","model_fit_plot","regional_plot","regional_rate_plot","linear_local_plot","non_linear_local_plot","non_linear_local_rate_plot","nigam_component_plot"}

\item{...}{Not used}
}
\value{
Plot of model fit and the rate of change depending on the statistical model in question.
}
\description{
For the EIV simple linear regression and the Change point models the model fit plots are created, preferably for 1 proxy site.
For the EIV IGP and the NI spline in time the plots of the model fits and the derivative of the model fit are produced for 1 proxy site, as running these models with more than 1 site or with tide gauge data is not recommended.
For the NI spline in space time, the plots of the model fits and the derivatives of the model fits are produced for any amount of proxy sites.
For the NI GAM decomposition, the plots of the model fits and the derivatives of the model fits are produced for any amount of sites including tide gauges and proxy sites. It is important to note that tide gauge data is strongly recommended for this model. Also, the ni_gam_decomp model will produce plots for each individual component,i.e. the regional component and its rate of change, the linear local component, the non-linear local component and its rate of change.
If tide gauges are used in the model, the user has the ability plot the output with or without the inclusion of the tide gauges.
}
\examples{
data <- NAACproxydata \%>\% dplyr::filter(Site == "Cedar Island")
x <- reslr_load(data = data)
jags_output <- reslr_mcmc(x, model_type = "eiv_slr_t")
plot(x =jags_output)
}
