% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.reslr_input.R
\name{print.reslr_input}
\alias{print.reslr_input}
\title{Print reslr input object}
\usage{
\method{print}{reslr_input}(x, ...)
}
\arguments{
\item{x}{An object of class \code{reslr_input}}

\item{...}{Other argunemnts (not supported)}
}
\value{
A neat presentation of your input reslr object
}
\description{
Print reslr input object
}
\examples{
data <- NAACproxydata \%>\% dplyr::filter(Site == "Cedar Island")
reslr_input <- reslr_load(data = data)
print.reslr_input(x = reslr_input)
}
