% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reslr_mcmc.R
\name{reslr_mcmc}
\alias{reslr_mcmc}
\title{Running a range of Bayesian statistical models using JAGS}
\usage{
reslr_mcmc(
  input_data,
  model_type,
  n_cp = 1,
  igp_smooth = 0.2,
  n_iterations = 5000,
  n_burnin = 1000,
  n_thin = 4,
  n_chains = 3,
  CI = 0.95,
  spline_nseg_t = NULL,
  spline_nseg_st = NULL
)
}
\arguments{
\item{input_data}{Input data from the reslr_load function}

\item{model_type}{Statistical Model type}

\item{n_cp}{Number of change points 1,2 or 3}

\item{igp_smooth}{Informs prior for the smoothness (correlation) parameter if model = "igp" is chosen. Choose a value between 0 and 1. Closer to 1 will increase smoothness.}

\item{n_iterations}{Number of iterations. Increasing this value will increase the computational run time}

\item{n_burnin}{Burnin value. This number removes a certain number of samples at the beginning.}

\item{n_thin}{Thinning value}

\item{n_chains}{Number of chains. The number of times the model will be run.}

\item{CI}{Size of the credible interval required by the user. The default is 0.95 corresponding to 95\%.}

\item{spline_nseg_t}{Number of segments required in a spline in time which is used the basis functions}

\item{spline_nseg_st}{Number of segments required in a spline in space time which is used the basis functions}

\item{xr}{Range of data from cross validation}

\item{xl}{Range of data for cross validation}
}
\value{
A list containing the input data, the JAGS output and output dataframes used for final plots. The output of this function is a list containing the input data, the JAGS output and output dataframes used for final plots.
}
\description{
In this function, a variety of statistical models can be run depending on the requirements of the user. All models are written within a Bayesian framework and use JAGS (Just Another Gibbs Sampler) to calculate Markov Chain Monte Carlo (MCMC) simulation to obtain estimates of unknown parameters. The user has the cability to alter the number of iterations, the number of burnin, the number of chains and the thinning. These options relate to the amount of MCMC simulations required and should be reviewed by the user to ensure model convergence is achieved without excessively long run times.
Prior to using this function, the user is recommended to use the reslr_load function which organises the input data into the correct structure required for this function.
The user chooses their 'model_type' and as a range of models to choose from. The simplest model available is the EIV simple linear regression. Followed by the EIV change point model which the user change decide how many change points are required using the n_cp function. This is limited to 3 change points.
The EIV IGP and NI spline in time models allows for model fits as well as rate of change estimates for one site as it is not recommended to use more than one site for all these models mentioned above.
The final two model options allow for examining model fits as well as rate of change estimates for multiple sites where the number of sites has no limit. The first option is the NI spline in space time.
The final option is the NI GAM decomposition which also allows for the decomposition of the signal into a regional, a linear local component and a non-linear local component as describe by Upton et al 2023.
}
\examples{
data <- NAACproxydata \%>\% dplyr::filter(Site == "Cedar Island")
input_data <- reslr_load(data = data)
reslr_mcmc(input_data = input_data, model_type = "eiv_slr_t")
}
