% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.reslr_output.R
\name{summary.reslr_output}
\alias{summary.reslr_output}
\title{Summarise the output created with \code{\link{reslr_mcmc}}
Produces summaries and convergence diagnostics for an object created with \code{\link{reslr_mcmc}}.
The different options are:
"diagnostics" to assess MCMC convergence}
\usage{
\method{summary}{reslr_output}(
  jags_output,
  type = c("diagnostics", "quantiles", "statistics", "parameter_estimate"),
  ...
)
}
\arguments{
\item{jags_output}{Output object from the \code{\link{reslr_mcmc}}}

\item{type}{User decides which type of summary they require}

\item{...}{Not in use}
}
\value{
A list containing convergence diagnostics
}
\description{
Summarise the output created with \code{\link{reslr_mcmc}}
Produces summaries and convergence diagnostics for an object created with \code{\link{reslr_mcmc}}.
The different options are:
"diagnostics" to assess MCMC convergence
}
\examples{
data <- NAACproxydata \%>\% dplyr::filter(Site == "Cedar Island")
input_data <- reslr_load(data = data)
jags_output  <- reslr_mcmc(input_data = input_data, model_type = "eiv_slr_t")
summary.reslr_output(jags_output = jags_output, type = "diagnostics")
}
